requirements.txt
elasticsearch>=2.4.0,<5.1.0
elasticsearch-curator>=4.1.0
argparse
json